Ventanas eliminar/ tablas datos en scrollpanes

+++ src/gui/JDEliminar.java
@@ -25,13 +25,25 @@ public class JDEliminar extends JDialog {
                     String codigoEquipo = datosRed.getJTEquipos().getModel().getValueAt(filaEquipo, 0).toString();
 
                     Equipo equipo = coordinador.buscarEquipo(codigoEquipo);
-                    try {
-                        coordinador.borrarEquipo(equipo);
-                        coordinador.cargarDatos();
-                        datosRed.actualizarTablaEquipos();
-                        JOptionPane.showMessageDialog(this, "Equipo eliminado exitosamente.");
-                    } catch (EquipoNoExistenteException | EquipoEnUsoException ENEEe) {
-                        JOptionPane.showMessageDialog(this, ENEEe.getMessage());
+                    int resultado = JOptionPane.showConfirmDialog(null ,"Seguro que quiere eliminar el Equipo: "
+                                    + "\nCodigo: " + equipo.getCodigo()
+                                    + "\nDescripcion: " + equipo.getDescripcion()
+                                    + "\nTipo de Equipo: " + equipo.getTipoEquipo().getCodigo() + ", " + equipo.getTipoEquipo().getDescripcion()
+                                    + "\nMarca: " + equipo.getMarca()
+                                    + "\nModelo: " + equipo.getModelo()
+                                    + "\nUbicacion: " + equipo.getUbicacion().getCodigo() + ", " + equipo.getUbicacion().getDescripcion()
+                            , "Eliminar Equipo"
+                            , JOptionPane.OK_CANCEL_OPTION, JOptionPane.PLAIN_MESSAGE);
+
+                    if (resultado == JOptionPane.OK_OPTION) {
+                        try {
+                            coordinador.borrarEquipo(equipo);
+                            coordinador.cargarDatos();
+                            datosRed.actualizarTablaEquipos();
+                            JOptionPane.showMessageDialog(this, "Equipo eliminado exitosamente.");
+                        } catch (EquipoNoExistenteException | EquipoEnUsoException ENEEe) {
+                            JOptionPane.showMessageDialog(this, ENEEe.getMessage());
+                        }
                     }
                 }
                 break;
@@ -45,13 +57,22 @@ public class JDEliminar extends JDialog {
                     String codigoTipoCable = datosRed.getJTConexiones().getModel().getValueAt(filaConexion,2).toString();
 
                     Conexion conexion = coordinador.buscarConexion(codigoEquipo1, codigoEquipo2, codigoTipoCable);
-                    try {
-                        coordinador.borrarConexion(conexion);
-                        coordinador.cargarDatos();
-                        datosRed.actualizarTablaConexiones();
-                        JOptionPane.showMessageDialog(this,"Conexion eliminada exitosamente.");
-                    } catch (ConexionNoExistenteException CNEe) {
-                        JOptionPane.showMessageDialog(this, CNEe.getMessage());
+                    int resultado = JOptionPane.showConfirmDialog(null , "Seguro que quiere eliminar la Conexion: "
+                                    + "\nEquipo 1: " + conexion.getEquipo1().getCodigo() + ", " + conexion.getEquipo1().getDescripcion()
+                                    + "\nEquipo 2: " + conexion.getEquipo2().getCodigo() + ", " + conexion.getEquipo2().getDescripcion()
+                                    + "\nCable: " + conexion.getTipoCable().getCodigo() + conexion.getTipoCable().getDescripcion()
+                            , "Eliminar Conexion"
+                            , JOptionPane.OK_CANCEL_OPTION, JOptionPane.PLAIN_MESSAGE);
+
+                    if (resultado == JOptionPane.OK_OPTION) {
+                        try {
+                            coordinador.borrarConexion(conexion);
+                            coordinador.cargarDatos();
+                            datosRed.actualizarTablaConexiones();
+                            JOptionPane.showMessageDialog(this, "Conexion eliminada exitosamente.");
+                        } catch (ConexionNoExistenteException CNEe) {
+                            JOptionPane.showMessageDialog(this, CNEe.getMessage());
+                        }
                     }
                 }
                 break;
@@ -61,13 +82,21 @@ public class JDEliminar extends JDialog {
                     String codigoUbicacion = datosRed.getJTUbicaciones().getModel().getValueAt(filaUbicacion, 0).toString();
 
                     Ubicacion ubicacion = coordinador.buscarUbicacion(codigoUbicacion);
-                    try {
-                        coordinador.borrarUbicacion(ubicacion);
-                        coordinador.cargarDatos();
-                        datosRed.actualizarTablaUbicaciones();
-                        JOptionPane.showMessageDialog(this, "Ubicacion eliminada exitosamente.");
-                    } catch (UbicacionNoExistenteException | UbicacionEnUsoException UNEEe) {
-                        JOptionPane.showMessageDialog(this, UNEEe.getMessage());
+                    int resultado = JOptionPane.showConfirmDialog(null , "Seguro que quiere eliminar la Ubicacion:"
+                                    + "\nCodigo: " + ubicacion.getCodigo()
+                                    + "\nDescripcion: " + ubicacion.getDescripcion()
+                            , "Eliminar Ubicacion"
+                            , JOptionPane.OK_CANCEL_OPTION, JOptionPane.PLAIN_MESSAGE);
+
+                    if (resultado == JOptionPane.OK_OPTION) {
+                        try {
+                            coordinador.borrarUbicacion(ubicacion);
+                            coordinador.cargarDatos();
+                            datosRed.actualizarTablaUbicaciones();
+                            JOptionPane.showMessageDialog(this, "Ubicacion eliminada exitosamente.");
+                        } catch (UbicacionNoExistenteException | UbicacionEnUsoException UNEEe) {
+                            JOptionPane.showMessageDialog(this, UNEEe.getMessage());
+                        }
                     }
                 }
                 break;
@@ -77,13 +106,21 @@ public class JDEliminar extends JDialog {
                     String codigoTipoEquipo = datosRed.getJTTipoEquipo().getModel().getValueAt(filaTipoEquipo, 0).toString();
 
                     TipoEquipo tipoEquipo = coordinador.buscarTipoEquipo(codigoTipoEquipo);
-                    try {
-                        coordinador.borrarTipoEquipo(tipoEquipo);
-                        coordinador.cargarDatos();
-                        datosRed.actualizarTablaTipoEquipos();
-                        JOptionPane.showMessageDialog(this, "Tipo de equipo eliminado exitosamente.");
-                    } catch (TipoEquipoEnUsoException TEEUEe) {
-                        JOptionPane.showMessageDialog(this, TEEUEe.getMessage());
+                    int resultado = JOptionPane.showConfirmDialog(null , "Seguro que quiere eliminar el Tipo de Equipo:"
+                                    + "\nCodigo: " + tipoEquipo.getCodigo()
+                                    + "\nDescripcion: " + tipoEquipo.getDescripcion()
+                            , "Eliminar Tipo de Equipo"
+                            , JOptionPane.OK_CANCEL_OPTION, JOptionPane.PLAIN_MESSAGE);
+
+                    if (resultado == JOptionPane.OK_OPTION) {
+                        try {
+                            coordinador.borrarTipoEquipo(tipoEquipo);
+                            coordinador.cargarDatos();
+                            datosRed.actualizarTablaTipoEquipos();
+                            JOptionPane.showMessageDialog(this, "Tipo de equipo eliminado exitosamente.");
+                        } catch (TipoEquipoEnUsoException TEEUEe) {
+                            JOptionPane.showMessageDialog(this, TEEUEe.getMessage());
+                        }
                     }
                 }
                 break;
@@ -93,13 +130,22 @@ public class JDEliminar extends JDialog {
                     String codigoTipoPuerto = datosRed.getJTTipoPuerto().getModel().getValueAt(filaTipoPuerto, 0).toString();
 
                     TipoPuerto tipoPuerto = coordinador.buscarTipoPuerto(codigoTipoPuerto);
-                    try {
-                        coordinador.borrarTipoPuerto(tipoPuerto);
-                        coordinador.cargarDatos();
-                        datosRed.actualizarTablaTipoPuertos();
-                        JOptionPane.showMessageDialog(this, "Tipo de Puerto eliminado exitosamente.");
-                    } catch (TipoPuertoEnUsoException TPEUEe) {
-                        JOptionPane.showMessageDialog(this, TPEUEe.getMessage());
+                    int resultado = JOptionPane.showConfirmDialog(null , "Seguro que quiere eliminar el Tipo de Puerto:"
+                                    + "\nCodigo: " + tipoPuerto.getCodigo()
+                                    + "\nDescripcion: " + tipoPuerto.getDescripcion()
+                                    + "\nVelocidad: " + tipoPuerto.getVelocidad()
+                            , "Eliminar Tipo de Puerto"
+                            , JOptionPane.OK_CANCEL_OPTION, JOptionPane.PLAIN_MESSAGE);
+
+                    if (resultado == JOptionPane.OK_OPTION) {
+                        try {
+                            coordinador.borrarTipoPuerto(tipoPuerto);
+                            coordinador.cargarDatos();
+                            datosRed.actualizarTablaTipoPuertos();
+                            JOptionPane.showMessageDialog(this, "Tipo de Puerto eliminado exitosamente.");
+                        } catch (TipoPuertoEnUsoException TPEUEe) {
+                            JOptionPane.showMessageDialog(this, TPEUEe.getMessage());
+                        }
                     }
                 }
                 break;
@@ -109,13 +155,22 @@ public class JDEliminar extends JDialog {
                     String codigoTipoCable = datosRed.getJTTipoCable().getModel().getValueAt(filaTipoCable,0).toString();
 
                     TipoCable tipoCable = coordinador.buscarTipoCable(codigoTipoCable);
-                    try {
-                        coordinador.borrarTipoCable(tipoCable);
-                        coordinador.cargarDatos();
-                        datosRed.actualizarTablaTipoCables();
-                        JOptionPane.showMessageDialog(this, "Tipo de cable eliminado exitosamente.");
-                    } catch (TipoCableEnUsoException TCEUEe) {
-                        JOptionPane.showMessageDialog(this, TCEUEe.getMessage());
+                    int resultado = JOptionPane.showConfirmDialog(null , "Seguro que quiere eliminar el Tipo de Cable:"
+                                    + "\nCodigo: " + tipoCable.getCodigo()
+                                    + "\nDescripcion: " + tipoCable.getDescripcion()
+                                    + "\nVelocidad: " + tipoCable.getVelocidad()
+                            , "Eliminar Tipo de Cable"
+                            , JOptionPane.OK_CANCEL_OPTION, JOptionPane.PLAIN_MESSAGE);
+
+                    if (resultado == JOptionPane.OK_OPTION) {
+                        try {
+                            coordinador.borrarTipoCable(tipoCable);
+                            coordinador.cargarDatos();
+                            datosRed.actualizarTablaTipoCables();
+                            JOptionPane.showMessageDialog(this, "Tipo de cable eliminado exitosamente.");
+                        } catch (TipoCableEnUsoException TCEUEe) {
+                            JOptionPane.showMessageDialog(this, TCEUEe.getMessage());
+                        }
                     }
                 }
                 break;
